# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Food Court QR Ordering Platform API
  version: 1.0.0-pilot-STALE
  description: |
    # ⚠ THIS DOCUMENT IS STALE. DO NOT GENERATE A CLIENT FROM IT.

    Superseded by `docs/prd.md` v7.0 and by the running code. It is kept only so
    the divergence is visible; every disagreement below resolves in favour of
    the implementation.

    The warning lives here, in the file itself, rather than only in the PRD.
    Somebody who opens this directly — to generate a client, to check a field
    name, to write an integration — would otherwise get no warning at all, and
    a stale contract read in good faith produces confidently wrong code.

    ## What is known to be wrong

    | Here | Actually |
    | --- | --- |
    | `/qr/{token}` returns a `table` object | It resolves to a FOOD COURT. The table concept was removed — PRD §2.1 |
    | Order status `COMPLETED` | `COLLECTED`. Cooked and handed over are different events — PRD §7.1 |
    | Payment status `SUCCESS` | `AUTHORIZED` and `CAPTURED` are separate states — PRD §8.1 |
    | Menu item `is_available` boolean | `status` / `availability` / `inventory_mode`, three concepts — PRD §6 |
    | `POST /kds/orders/{id}/complete` | `/collect` |
    | no session or OTP endpoints | `/auth/otp/request`, `/auth/otp/verify` exist — PRD §11.2 |
    | no payment endpoints | `/orders/{id}/payment-intent`, `/webhooks/payments`, `/orders/{id}/payment` exist |
...

...
    script produces types that contradict the server.

...

...

...

...

...
servers:
...
security:
  - sessionToken: []
paths:
...
  /payments/{paymentId}:
    get:
      tags: [Payment]
      operationId: getPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Payment state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }

  # ==================================================================
  # ORDER — CUSTOMER
  # ==================================================================
...
  /webhooks/payments/{provider}:
    post:
      tags: [Webhooks]
      operationId: paymentWebhook
      security: []
      summary: Inbound payment provider callback
      description: |
        PRD PAY-03/04. **Signature is verified before the body is parsed.**
        Idempotent on the provider event id via a unique constraint on
        `processed_event`; a duplicate is acknowledged with `200` and dropped.

        A verified success event marks the payment `SUCCESS` and transitions the
        order to `PAYMENT_CONFIRMED` in one transaction, then enqueues dispatch
        as a separate committed command — so a webhook delivered twice results
        in exactly one dispatch.

        An unrecognised or unparseable event is **never** mapped to `FAILED`. It
        opens a `RECONCILIATION_REQUIRED` item.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [razorpay, cashfree, payu] }
        - name: X-Provider-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Provider-specific envelope. Shape is validated per adapter.
              additionalProperties: true
      responses:
        '200':
          description: Accepted, or recognised as a duplicate and dropped.
        '401':
          description: Signature verification failed. Logged and alerted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer session token issued by POST /sessions.
    deviceToken:
      type: http
      scheme: bearer
      description: KDS or printer device token. Scoped to one vendor.
    staffToken:
      type: http
      scheme: bearer
      description: Manager, ops, finance and admin. Role and tenant scope are claims.
...
  schemas:
    Problem:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          description: |
            Stable machine-readable error code. Clients branch on this, never on
            the message. See TDD §18 for the full catalogue.
          examples:
            [VENDOR_DEVICE_OFFLINE, INVALID_TRANSITION, QUOTE_EXPIRED, IDEMPOTENCY_KEY_REQUIRED]
        message: { type: string, description: Human-readable, not for display to customers. }
        detail: { type: string, nullable: true }
        correlationId: { type: string }
        fields:
          type: object
          nullable: true
          additionalProperties: { type: string }
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Amount in paise. 25000 = ₹250.00. Never a decimal.
...
    PaymentStatus:
      type: string
      enum:
        [
          INITIATED,
          PENDING,
          SUCCESS,
          FAILED,
          REFUND_PENDING,
          REFUNDED,
          REFUND_FAILED,
          RECONCILIATION_REQUIRED,
        ]
...
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        method: { type: string, nullable: true }
        amountPaise: { $ref: '#/components/schemas/Money' }
        createdAt: { type: string, format: date-time }
        succeededAt: { type: string, format: date-time, nullable: true }
...